Some return type and param changes (see ext. description)
Because the Display class doesn't have access to the gameArray in GameCheckDriver, many of the methods in GameCheckDriver should have an ArrayList<Game> as a parameter (for searching/sorting) and an ArrayList<Game> as a return type (to send games to the Display class to be presented in the UI). These methods should also be static, because for the purposes of our program, there is no such thing as a GameCheckDriver object, and you cannot make a static reference to a non-static method (epic Java moment). I have implemented these changes.